Extend test coverage. Minor refactoring.
